core&ui: add contest private files
