+++ packages/ui-default/pages/contest.page.styl
@@ -217,5 +217,7 @@ $highlight-button-color = #F6DF45
     width: rem(130px)
 
 .page--contest_manage
+  .col--id
+    width: rem(80px)
   .col--score
     width: rem(80px)
