+++ packages/ui-default/templates/partials/contest_sidebar_management.html
@@ -29,6 +29,9 @@
       <li class="menu__item"><a class="menu__link" href="{{ url('contest_balloon', tid=tdoc.docId) }}">
         <span class="icon icon-balloon"></span> {{ _('Balloon Status') }}
       </a></li>
+      <li class="menu__item"><a class="menu__link" href="{{ url('contest_clarification', tid=tdoc.docId) }}">
+        <span class="icon icon-comment--text"></span> {{ _('Clarification') }}
+      </a></li>
       <li class="menu__seperator"></li>
     </ol>
   </div>
